Refactor keyboard usage monitor + smoother zooming on photo editor
